ratatui_crossterm

Struct CrosstermBackend

pub struct CrosstermBackend<W: Write> { /* private fields */ }

A Backend implementation that uses Crossterm to render to the terminal.

The CrosstermBackend struct is a wrapper around a writer implementing Write, which is used to send commands to the terminal. It provides methods for drawing content, manipulating the cursor, and clearing the terminal screen.

Most applications should not call the methods on CrosstermBackend directly, but will instead use the Terminal struct, which provides a more ergonomic interface.

Usually applications will enable raw mode and switch to alternate screen mode after creating a CrosstermBackend. This is done by calling [crossterm::terminal::enable_raw_mode] and [crossterm::terminal::EnterAlternateScreen] (and the corresponding disable/leave functions when the application exits). This is not done automatically by the backend because it is possible that the application may want to use the terminal for other purposes (like showing help text) before entering alternate screen mode.

Example

use std::io::{stderr, stdout};

use crossterm::ExecutableCommand;
use crossterm::terminal::{
    EnterAlternateScreen, LeaveAlternateScreen, disable_raw_mode, enable_raw_mode,
};
use ratatui::Terminal;
use ratatui::backend::CrosstermBackend;

let mut backend = CrosstermBackend::new(stdout());
// or
let backend = CrosstermBackend::new(stderr());
let mut terminal = Terminal::new(backend)?;

enable_raw_mode()?;
stdout().execute(EnterAlternateScreen)?;

terminal.clear()?;
terminal.draw(|frame| {
    // -- snip --
})?;

stdout().execute(LeaveAlternateScreen)?;
disable_raw_mode()?;

See the the Examples directory for more examples. See the backend module documentation for more details on raw mode and alternate screen.

Implementations

impl<W> CrosstermBackend<W>
where W: Write,

pub const fn new(writer: W) -> Self

Creates a new CrosstermBackend with the given writer.

Most applications will use either stdout or stderr as writer. See the FAQ to determine which one to use.

Example

use std::io::stdout;

use ratatui::backend::CrosstermBackend;

let backend = CrosstermBackend::new(stdout());

pub const fn writer(&self) -> &W

Gets the writer.

Stability

This API is marked as unstable and is only available when the unstable-backend-writer crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time. The tracking issue is: https://github.com/ratatui/ratatui/pull/991.

pub const fn writer_mut(&mut self) -> &mut W

Gets the writer as a mutable reference.

Note: writing to the writer may cause incorrect output after the write. This is due to the way that the Terminal implements diffing Buffers.

Stability

This API is marked as unstable and is only available when the unstable-backend-writer crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time. The tracking issue is: https://github.com/ratatui/ratatui/pull/991.