ratatui_crossterm

Trait FromCrossterm

pub trait FromCrossterm<C> {
    // Required method
    fn from_crossterm(value: C) -> Self;
}

A trait for converting a Crossterm type to a Ratatui type.

This trait is needed for avoiding the orphan rule when implementing From for crossterm types once these are moved to a separate crate.

Required Methods

fn from_crossterm(value: C) -> Self

Converts the crossterm type to a ratatui type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.