ratatui_crossterm

Trait IntoCrossterm

pub trait IntoCrossterm<C> {
    // Required method
    fn into_crossterm(self) -> C;
}

A trait for converting a Ratatui type to a Crossterm type.

This trait is needed for avoiding the orphan rule when implementing From for crossterm types once these are moved to a separate crate.

Required Methods

fn into_crossterm(self) -> C

Converts the ratatui type to a crossterm type.